implement RFC81 label-leader offsetting heavy refactoring in msDrawLabelCache and msDrawShape to cut down on processing time and memory consumption
git-svn-id: http://svn.osgeo.org/mapserver/trunk@13224 7532c77e-422f-0410-93f4-f0b67bdd69e2

+++ maprendering.c
@@ -812,39 +812,41 @@ int msDrawMarkerSymbol(symbolSetObj *symbolset,imageObj *image, pointObj *p, sty
          symbolObj *symbol = symbolset->symbol[style->symbol];
          /* store a reference to the renderer to be used for freeing */
          symbol->renderer = renderer;
-         switch (symbol->type) 
-	 {
-	   case (MS_SYMBOL_TRUETYPE): {
-	     if(!symbol->full_font_path)
-               symbol->full_font_path =  msStrdup(msLookupHashTable(&(symbolset->fontset->fonts),
-								    symbol->font));
-	     if(!symbol->full_font_path) {
-               msSetError(MS_MEMERR,"allocation error", "msDrawMarkerSymbol()");
-               return MS_FAILURE;
-	     }
-	   }
-	     break;
-	   case (MS_SYMBOL_PIXMAP): {
-	     if(!symbol->pixmap_buffer) {
-               if(MS_SUCCESS != msPreloadImageSymbol(renderer,symbol))
-		 return MS_FAILURE;
-	     }
-	   }
-	     break;
-         
-	   case (MS_SYMBOL_SVG): {
+         switch (symbol->type) {
+	         case (MS_SYMBOL_TRUETYPE):
+            {
+               if (!symbol->full_font_path)
+                  symbol->full_font_path = msStrdup(msLookupHashTable(&(symbolset->fontset->fonts),
+                       symbol->font));
+               if (!symbol->full_font_path) {
+                  msSetError(MS_MEMERR, "allocation error", "msDrawMarkerSymbol()");
+                  return MS_FAILURE;
+               }
+            }
+            break;
+            case (MS_SYMBOL_PIXMAP):
+            {
+               if (!symbol->pixmap_buffer) {
+                  if (MS_SUCCESS != msPreloadImageSymbol(renderer, symbol))
+                     return MS_FAILURE;
+               }
+            }
+            break;
+
+            case (MS_SYMBOL_SVG):
+            {
 #ifdef USE_SVG_CAIRO
-	     if(!symbol->svg_cairo_surface) {
-	       if(MS_SUCCESS != msPreloadSVGSymbol(symbol))
-		 return MS_FAILURE;
-	     }
+               if (!symbol->svg_cairo_surface) {
+                  if (MS_SUCCESS != msPreloadSVGSymbol(symbol))
+                     return MS_FAILURE;
+               }
 #else
-	     msSetError(MS_SYMERR, "SVG symbol support is not enabled.", "msDrawMarkerSymbol()");
-	     return MS_FAILURE;
+               msSetError(MS_SYMERR, "SVG symbol support is not enabled.", "msDrawMarkerSymbol()");
+               return MS_FAILURE;
 #endif
-	   }
-	   break;
-	 }
+            }
+            break;
+         }
 
          s.style = style;
          computeSymbolStyle(&s,style,symbol,scalefactor);
@@ -862,32 +864,31 @@ int msDrawMarkerSymbol(symbolSetObj *symbolset,imageObj *image, pointObj *p, sty
          	return; // size too small
           */
 
-	 p_x = p->x;
-	 p_y = p->y;
+         p_x = p->x;
+         p_y = p->y;
 
-	 if (style->polaroffsetpixel != 0 ||
-	     style->polaroffsetangle != 0) 
-	 {
-	   double angle = style->polaroffsetangle * MS_DEG_TO_RAD;
-	   p_x +=  (style->polaroffsetpixel * cos(-angle)) * scalefactor; 
-	   p_y +=  (style->polaroffsetpixel * sin(-angle)) * scalefactor;
-	 }
+         if (style->polaroffsetpixel != 0 ||
+               style->polaroffsetangle != 0) 
+         {
+            double angle = style->polaroffsetangle * MS_DEG_TO_RAD;
+            p_x +=  (style->polaroffsetpixel * cos(-angle)) * scalefactor; 
+            p_y +=  (style->polaroffsetpixel * sin(-angle)) * scalefactor;
+         }
 
-	 p_x +=  style->offsetx * scalefactor;
-	 p_y +=  style->offsety * scalefactor;
+         p_x +=  style->offsetx * scalefactor;
+         p_y +=  style->offsety * scalefactor;
 
          if(symbol->anchorpoint_x != 0.5 || symbol->anchorpoint_y != 0.5) {
-            int sx,sy;
+            double sx,sy;
             double ox, oy;
             msGetMarkerSize(symbolset, style, &sx, &sy, scalefactor);
             ox = (0.5 - symbol->anchorpoint_x) * sx;
             oy = (0.5 - symbol->anchorpoint_y) * sy;
             if(s.rotation != 0) {
-	       double sina, cosa;
+               double sina, cosa;
                double rox,roy;
-	       /* sincos(-s.rotation,&sina,&cosa); */
-	       sina = sin(-s.rotation);
-	       cosa = cos(-s.rotation);
+               sina = sin(-s.rotation);
+               cosa = cos(-s.rotation);
                rox = ox * cosa - oy * sina;
                roy = ox * sina + oy * cosa;
                p_x += rox;
